+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   take a new `HexOrientation` parameter (#189)
 * (**BREAKING**) `HexLayout` Y axis is no longer inverted by default (#187)
 * `HexLayout` builder patter (#187)
+* (**BREAKING**) `HexLayout` removed the `invert_x` and `invert_y` fields (#189)
+* (**BREAKING**) `HexLayout` `hex_size` field is now `scale` (#189)
 
 ## 0.19.1